pkg/server: isolate initstats tests into dedicated package
