+++ pkg/server/tests/initstats/BUILD.bazel
@@ -1,0 +1,27 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "initstats_test",
+    timeout = "short",
+    srcs = [
+        "initstats_test.go",
+        "main_test.go",
+    ],
+    flaky = True,
+    shard_count = 2,
+    deps = [
+        "//pkg/config",
+        "//pkg/metrics",
+        "//pkg/server",
+        "//pkg/session",
+        "//pkg/store/mockstore",
+        "//pkg/store/mockstore/unistore",
+        "//pkg/testkit/testsetup",
+        "//pkg/util/topsql/state",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//oracle",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
